requirements/pip.txt
numpy
matplotlib
diffpy.utils
